+++ tests/no_version/test_auth_groups_api.py
@@ -73,7 +73,7 @@ async def test_update_group(registered_superuser_client):
     assert res.status_code == 200
     assert res.json()["user_ids"] == []
 
-    # Patch an existing group by adding a valid users
+    # Patch an existing group by adding a valid user
     res = await registered_superuser_client.patch(
         f"{PREFIX}/group/{group_id}/",
         json=dict(new_user_ids=[user_A_id]),
@@ -150,6 +150,31 @@ async def test_user_group_crud(registered_superuser_client):
     assert len(groups_data) == 2
     for group in groups_data:
         assert group["user_ids"] is None
+
+    # Add users A and B to group 2, and fail because user B is already there
+    res = await registered_superuser_client.patch(
+        f"{PREFIX}/group/{group_2_id}/",
+        json=dict(new_user_ids=[user_A_id, user_B_id]),
+    )
+    assert res.status_code == 422
+    hint_msg = "Likely reason: one of these links already exists"
+    assert hint_msg in res.json()["detail"]
+
+    # After the previous 422, verify that user A was not added to group 2
+    # (that is, verify that `db.commit` is atomic)
+    res = await registered_superuser_client.get(
+        f"{PREFIX}/group/{group_2_id}/"
+    )
+    assert res.status_code == 200
+    assert user_A_id not in res.json()["user_ids"]
+
+    # Create user/group link and fail because of repeated IDs
+    res = await registered_superuser_client.patch(
+        f"{PREFIX}/group/{group_1_id}/",
+        json=dict(new_user_ids=[99, 99]),
+    )
+    assert res.status_code == 422
+    assert "`new_user_ids` list has repetitions'" in str(res.json())
 
 
 async def test_get_user_group_names(